Lab04 Code

+++ Facultate An 3/VVSS/2-ProiectSalariati/ProiectSalariati/src/main/java/salariati/model/Employee.java
@@ -105,7 +105,7 @@ public class Employee {
     /**
      * Getter for the employee salary
      */
-    public float getSalary() {
+    public int getSalary() {
         return salary;
     }
